# backend/openapi.yaml
openapi: 3.0.3
info:
    title: Cardex API
    version: 1.0.0
    description: |
        API for the Cardex trading card collection application.

        Card data is sourced from the [Pokémon TCG API](https://api.pokemontcg.io/v2).
        User collection and wishlist data is stored in Supabase.

        User card endpoints require a Supabase access token in the `Authorization` header.
servers:
    - url: /
      description: API server
...
paths:
...
    /v1/user-cards:
        post:
            tags:
                - User Cards
            summary: Add card to collection or wishlist
            description: |
                Adds a card to the user's owned collection or wishlist.
                When `card_name` is included (e.g. from search results), the backend
                stores that snapshot in Supabase without calling the Pokémon TCG API.
                If `card_name` is omitted, the backend fetches the card upstream first.
            operationId: createUserCard
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserCardRequest'
                        examples:
                            owned:
                                summary: Owned card
                                value:
                                    external_card_id: swsh11-131
                                    status: owned
                                    quantity: 1
                                    condition: near_mint
                                    notes: Pulled from a pack
                                    card_name: Giratina V
                                    card_image_url: https://images.pokemontcg.io/swsh11/131_hires.png
                                    market_price: 12.5
                            wishlist:
                                summary: Wishlist card
                                value:
                                    external_card_id: swsh11-131
                                    status: wishlist
                                    notes: Want this card later
                                    card_name: Giratina V
                                    card_image_url: https://images.pokemontcg.io/swsh11/131_hires.png
                                    market_price: 12.5
            responses:
                '201':
                    description: Card added successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserCard'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '401':
                    description: Authentication required
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: Sign in to save cards.
                '404':
                    description: Card not found in the Pokémon TCG API
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: We could not find that card. Try searching again.
                '409':
                    description: Card already saved with this status
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            examples:
                                owned:
                                    summary: Already in collection
                                    value:
                                        error: This card is already in your collection.
                                wishlist:
                                    summary: Already on wishlist
                                    value:
                                        error: This card is already on your wishlist.
                '502':
                    description: Upstream Pokémon TCG API error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: Unable to look up card details right now. Please try again.
                '504':
                    description: Upstream Pokémon TCG API timed out
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: Card lookup timed out. Please try again.
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: Unable to save the card. Please try again.
    /v1/user-cards/me:
        get:
            tags:
                - User Cards
            summary: Get current user's cards
            description: |
                Returns the authenticated user's saved cards, newest first.
                Optionally filter by `status`.
            operationId: getMyUserCards
            security:
                - bearerAuth: []
            parameters:
                - name: status
                  in: query
                  required: false
                  description: Filter by owned or wishlist cards
                  schema:
                      $ref: '#/components/schemas/CardStatus'
            responses:
                '200':
                    description: List of user cards
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/UserCard'
                '400':
                    description: Invalid status filter
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '401':
                    description: Authentication required
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /v1/user-cards/{id}:
        patch:
            tags:
                - User Cards
            summary: Update a user card
            description: Updates one or more fields on a user card owned by the authenticated user.
            operationId: updateUserCard
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  description: User card row ID
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserCardRequest'
            responses:
                '200':
                    description: Updated user card
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserCard'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '401':
                    description: Authentication required
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: User card not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: User card not found
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
        delete:
            tags:
                - User Cards
            summary: Remove a user card
            description: Deletes a card from the user's collection or wishlist.
            operationId: deleteUserCard
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  description: User card row ID
                  schema:
                      type: string
                      format: uuid
            responses:
                '204':
                    description: User card deleted
                '401':
                    description: Authentication required
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: User card not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Supabase access token from sign-in
    schemas:
        ErrorResponse:
            type: object
            required:
                - error
            properties:
                error:
                    type: string
                    example: Something went wrong
...
        CardCondition:
            type: string
            enum:
                - mint
                - near_mint
                - lightly_played
                - moderately_played
                - heavily_played
                - damaged
        CardStatus:
            type: string
            enum:
                - owned
                - wishlist
...
        UserCard:
            type: object
            required:
                - id
                - user_id
                - external_card_id
                - status
                - created_at
                - updated_at
            properties:
                id:
                    type: string
                    format: uuid
                user_id:
                    type: string
                    format: uuid
                external_card_id:
                    type: string
                    example: swsh11-131
                status:
                    $ref: '#/components/schemas/CardStatus'
                quantity:
                    type: integer
                    nullable: true
                    example: 1
                condition:
                    allOf:
                        - $ref: '#/components/schemas/CardCondition'
                    nullable: true
                notes:
                    type: string
                    nullable: true
                card_name:
                    type: string
                    nullable: true
                    example: Giratina V
                card_image_url:
                    type: string
                    format: uri
                    nullable: true
                market_price:
                    type: number
                    nullable: true
                    description: Cached TCGPlayer market price in USD at time of save
                    example: 8.01
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
        CreateUserCardRequest:
            type: object
            required:
                - external_card_id
                - status
            properties:
                external_card_id:
                    type: string
                    example: swsh11-131
                status:
                    $ref: '#/components/schemas/CardStatus'
                quantity:
                    type: integer
                    description: Defaults to 1 when status is owned
                    example: 1
                condition:
                    $ref: '#/components/schemas/CardCondition'
                notes:
                    type: string
                card_name:
                    type: string
                    description: |
                        Display name from search results. When provided, skips an upstream
                        Pokémon TCG API lookup on add.
                card_image_url:
                    type: string
                    nullable: true
                market_price:
                    type: number
                    nullable: true
        UpdateUserCardRequest:
            type: object
            properties:
                status:
                    $ref: '#/components/schemas/CardStatus'
                quantity:
                    type: integer
                condition:
                    allOf:
                        - $ref: '#/components/schemas/CardCondition'
                    nullable: true
                notes:
                    type: string
                    nullable: true